HashSet takes the throne First change : All ArrayList are replaced with HashSet in classes (Rule, ExpertSystem) and all necessary changes were made. HashSet have a complexity of O(1) in adding, deleting, contains methods, and since we do not care about the order of statement there is no point in using an ArrayList.
Second change : Adding a static function in Statement class called addTo(HashSet<Statement> statements, Statement s). This functions adds a new statement 's' to 'statements' while preserving the non-redundancy of 'statements'. This function assume that 'statements' is not redundant in the first place. This function is used in 'addKnowledge' in ExpertSystem class, addAntecedant and addConsequence in Rule class. If we build a set of statements with this function we can be sure that the result will be non-redundant.

Third change : Recoding the 'infer' function in ExpertSystem class. The goal was to optimize it with the new implemented HashSet and other optimization that was talked before.